requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
blis==1.2.0
catalogue==2.0.10
celery==5.4.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.21.0
confection==0.1.5
cymem==2.0.11
Django==5.1.7
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docx==0.2.4
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
joblib==1.4.2
kombu==5.5.1
langcodes==3.5.0
language_data==1.3.0
language_tool_python==2.9.0
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
nltk==3.9.1
numpy==2.2.4
packaging==24.2
pika==1.3.2
pillow==11.1.0
preshed==3.0.9
prompt_toolkit==3.0.50
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rest-framework-simplejwt==0.0.2
rich==13.9.4
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
thinc==8.3.4
toml==0.10.2
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.2
